clk_wiz: Corrected the type casting error in clk_wiz
Calculation of input and output frequencies is not giving float
output because of type casting error. Corrected the same by
type casting the numerator in the divison to float.

Acked-by: Vishal Sagar <vishal.sagar@xilinx.com>

Acked-by: Vishal Sagar <vishal.sagar@xilinx.com>

+++ XilinxProcessorIPLib/drivers/clk_wiz/examples/xclk_wiz_intr_example.c
@@ -294,7 +294,7 @@ int Clk_Wiz_Reconfig(XClk_Wiz_Config *CfgPtr_Dynamic)
     }
 
     /* Calculation of Input Freq and Divide factors*/
-    Freq = (CLK_WIZ_VCO_FACTOR/DYNAMIC_INPUT_FREQ_FACTOR);
+    Freq = ((float) CLK_WIZ_VCO_FACTOR/ DYNAMIC_INPUT_FREQ_FACTOR);
 
     Divide = Freq;
     Freq = (float)(Freq - Divide);
@@ -325,7 +325,7 @@ int Clk_Wiz_Reconfig(XClk_Wiz_Config *CfgPtr_Dynamic)
     *(u32 *)(CfgPtr_Dynamic->BaseAddr + 0x204) = 0x00;
 
     /* Calculation of Output Freq and Divide factors*/
-    Freq = (CLK_WIZ_VCO_FACTOR / DYNAMIC_OUTPUT_FREQFACTOR);
+    Freq = ((float) CLK_WIZ_VCO_FACTOR / DYNAMIC_OUTPUT_FREQFACTOR);
 
     Divide = Freq;
     Freq = (float)(Freq - Divide);
